% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_functions.R
\name{combine_person_data}
\alias{combine_person_data}
\title{Combine person-level data}
\usage{
combine_person_data(person_data)
}
\arguments{
\item{person_data}{A list of person-level datasets with the following columns: locus, allele, colour, n}
}
\description{
Combine multiple individual-person datasets into a long dataset
}
