% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ProfileManager.R
\name{ProfileManager}
\alias{ProfileManager}
\title{Profile Manager}
\description{
An R6 class that handles fetching STRBase tables and linking it with profile data
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{multiplex}}{The multiplex to be used}

\item{\code{fetched}}{List of fetched loci}

\item{\code{profile_data}}{Profile data, stored as a long dataframe (locus, colour, allele, count)}

\item{\code{unique_loci}}{Vector of unique loci present in the data}

\item{\code{invalid_loci}}{Vector of loci to skip during fetching and data cleaning}

\item{\code{tables}}{A list of all STRBase tables, one element per each unique (valid) locus}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-ProfileManager-new}{\code{ProfileManager$new()}}
\item \href{#method-ProfileManager-fetch_strbase_tables}{\code{ProfileManager$fetch_strbase_tables()}}
\item \href{#method-ProfileManager-extract_basepairs}{\code{ProfileManager$extract_basepairs()}}
\item \href{#method-ProfileManager-profile_plot}{\code{ProfileManager$profile_plot()}}
\item \href{#method-ProfileManager-clone}{\code{ProfileManager$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ProfileManager-new"></a>}}
\if{latex}{\out{\hypertarget{method-ProfileManager-new}{}}}
\subsection{Method \code{new()}}{
Create a new ProfileManager object
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ProfileManager$new(profile_data, multiplex = "PowerPlex 16")}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{profile_data}}{Profile data, in a long format}

\item{\code{multiplex}}{The multiplex to be used for linking}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ProfileManager-fetch_strbase_tables"></a>}}
\if{latex}{\out{\hypertarget{method-ProfileManager-fetch_strbase_tables}{}}}
\subsection{Method \code{fetch_strbase_tables()}}{
Retrives multiple sets of tables with information about gene loci from STRBase
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ProfileManager$fetch_strbase_tables(loci = NULL, n = NULL, refetch = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{loci}}{Loci to be retrived (defaults to all unique loci in the profile data)}

\item{\code{n}}{(Optional) fetch only the first n loci}

\item{\code{refetch}}{Refetch tables that have been already fetched}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ProfileManager-extract_basepairs"></a>}}
\if{latex}{\out{\hypertarget{method-ProfileManager-extract_basepairs}{}}}
\subsection{Method \code{extract_basepairs()}}{
Extract basepair information from STRBase tables
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ProfileManager$extract_basepairs(multiplex = "PowerPlex 16")}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{multiplex}}{The multiplex to be used for matching base-pairs}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ProfileManager-profile_plot"></a>}}
\if{latex}{\out{\hypertarget{method-ProfileManager-profile_plot}{}}}
\subsection{Method \code{profile_plot()}}{
Plot base pair information
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ProfileManager$profile_plot(
  col = c(blue = "#377eb8", green = "#4daf4a", yellow = "black")
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{col}}{Named vector of colours matching the colour variable in the dataset}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ProfileManager-clone"></a>}}
\if{latex}{\out{\hypertarget{method-ProfileManager-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ProfileManager$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
