% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ProfileManager.R
\name{ProfileManager}
\alias{ProfileManager}
\title{Profile Manager}
\description{
An R6 class that handles fetching STRBase tables and linking it with profile data
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{multiplex}}{The multiplex to be used}

\item{\code{profile_data}}{Profile data, stored as a long dataframe (locus, colour, allele, count)}

\item{\code{unique_loci}}{Vector of unique loci present in the data}

\item{\code{invalid_loci}}{Vector of loci to skip during fetching and data cleaning}

\item{\code{tables}}{A list of all STRBase tables, one element per each unique (valid) locus}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-ProfileManager-new}{\code{ProfileManager$new()}}
\item \href{#method-ProfileManager-fetch_strbase_tables}{\code{ProfileManager$fetch_strbase_tables()}}
\item \href{#method-ProfileManager-extract_basepairs}{\code{ProfileManager$extract_basepairs()}}
\item \href{#method-ProfileManager-clone}{\code{ProfileManager$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ProfileManager-new"></a>}}
\if{latex}{\out{\hypertarget{method-ProfileManager-new}{}}}
\subsection{Method \code{new()}}{
Create a new ProfileManager object
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ProfileManager$new(profile_data, multiplex = "Powerplex 16")}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{profile_data}}{Profile data, in a long format}

\item{\code{multiplex}}{The multiplex to be used for linking}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ProfileManager-fetch_strbase_tables"></a>}}
\if{latex}{\out{\hypertarget{method-ProfileManager-fetch_strbase_tables}{}}}
\subsection{Method \code{fetch_strbase_tables()}}{
Retrives multiple sets of tables with information about gene loci from STRBase
(one for each unique gene locus in the data)
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ProfileManager$fetch_strbase_tables(loci = NULL)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ProfileManager-extract_basepairs"></a>}}
\if{latex}{\out{\hypertarget{method-ProfileManager-extract_basepairs}{}}}
\subsection{Method \code{extract_basepairs()}}{
Extract basepair information from STRBase tables
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ProfileManager$extract_basepairs()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ProfileManager-clone"></a>}}
\if{latex}{\out{\hypertarget{method-ProfileManager-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ProfileManager$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
